--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,55 +1,33 @@
 package hw10programoptimization
 
 import (
-	"fmt"
+	"bufio"
 	"io"
-	"io/ioutil"
 	"strings"
 )
 
 type User struct {
-	ID       int
-	Name     string
-	Username string
-	Email    string
-	Phone    string
-	Password string
-	Address  string
+	Email string
 }
 
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	if err != nil {
-		return nil, fmt.Errorf("get users error: %w", err)
-	}
-	return countDomains(u, domain)
-}
-
-type users [100_000]User
-
-func getUsers(r io.Reader) (result users, err error) {
-	content, err := ioutil.ReadAll(r)
-	if err != nil {
-		return
-	}
-
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
-		var user User
-		if err = user.UnmarshalJSON([]byte(line)); err != nil {
-			return
-		}
-		result[i] = user
-	}
-	return
-}
-
-func countDomains(u users, domain string) (DomainStat, error) {
 	result := make(DomainStat)
 
-	for _, user := range u {
+	reader := bufio.NewReader(r)
+
+	for {
+		line, _, err := reader.ReadLine()
+
+		if err == io.EOF {
+			break
+		}
+
+		var user User
+		if err := user.UnmarshalJSON([]byte(line)); err != nil {
+			return nil, err
+		}
 
 		if strings.HasSuffix(user.Email, domain) {
 			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]++