--- a/hw10_program_optimization/stats_optimization_test.go
+++ b/hw10_program_optimization/stats_optimization_test.go
@@ -1,9 +1,11 @@
+//go:build bench
 // +build bench
 
 package hw10programoptimization
 
 import (
 	"archive/zip"
+	"bytes"
 	"testing"
 	"time"
 
@@ -25,7 +27,9 @@ func TestGetDomainStat_Time_And_Memory(t *testing.T) {
 
 		r, err := zip.OpenReader("testdata/users.dat.zip")
 		require.NoError(t, err)
-		defer r.Close()
+		defer func(r *zip.ReadCloser) {
+			_ = r.Close()
+		}(r)
 
 		require.Equal(t, 1, len(r.File))
 
@@ -47,6 +51,18 @@ func TestGetDomainStat_Time_And_Memory(t *testing.T) {
 
 	require.Less(t, int64(result.T), int64(timeLimit), "the program is too slow")
 	require.Less(t, mem, memoryLimit, "the program is too greedy")
+}
+
+func BenchmarkGetDomainStat(b *testing.B) {
+	data := `{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@Browsedrive.gov","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}
+{"Id":2,"Name":"Jesse Vasquez","Username":"qRichardson","Email":"mLynch@broWsecat.com","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}
+{"Id":3,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
+{"Id":4,"Name":"Gregory Reid","Username":"tButler","Email":"5Moore@Teklist.net","Phone":"520-04-16","Password":"r639qLNu","Address":"Sunfield Park 20"}
+{"Id":5,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@Linktype.com","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}`
+	bufferString := bytes.NewBufferString(data)
+	for i := 0; i < b.N; i++ {
+		_, _ = GetDomainStat(bufferString, "com")
+	}
 }
 
 var expectedBizStat = DomainStat{