## Домашнее задание №10 «Оптимизация программы»

### Чек-лист студента ([Что это?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0))
- [x] Я перечитал задание после решения. [Зачем перечитывать задание?](https://github.com/OtusGolang/home_work/wiki/%D0%9A%D0%BE%D0%BC%D0%BC%D0%B5%D0%BD%D1%82%D0%B0%D1%80%D0%B8%D0%B8-%D0%BA-%D1%87%D0%B5%D0%BA-%D0%BB%D0%B8%D1%81%D1%82%D1%83-%D1%81%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0#user-content-%D0%97%D0%B0%D1%87%D0%B5%D0%BC-%D0%BF%D0%B5%D1%80%D0%B5%D1%87%D0%B8%D1%82%D1%8B%D0%B2%D0%B0%D1%82%D1%8C-%D0%B7%D0%B0%D0%B4%D0%B0%D0%BD%D0%B8%D0%B5)
- [x] Я запустил `go mod tidy`.
- [x] Я удалил `.sync` файл. [Зачем его удалять?](https://github.com/OtusGolang/home_work/wiki/%5B%D0%A1%D1%82%D1%83%D0%B4%D0%B5%D0%BD%D1%82%D0%B0%D0%BC%5D-%D0%9F%D1%80%D0%BE%D1%86%D0%B5%D1%81%D1%81-%D1%81%D0%B4%D0%B0%D1%87%D0%B8-%D0%94%D0%97#user-content-%D0%92%D0%B0%D1%80%D0%B8%D0%B0%D0%BD%D1%82-2)
- [x] Я написал бенчмарк и приложил результаты benchstat.

### Критерии оценки
- [ ] Пайплайн зелёный и нет попытки «обмануть» систему - 4 балла
- [ ] Добавлены юнит-тесты - до 3 баллов
- [ ] Понятность и чистота кода - до 3 баллов

#### Зачёт от 7 баллов


Финальный результат бенчмарка. Промежутные результаты записаны внутри коммитов.

```
$ go test -v -count=1 -timeout=30s -tags bench .
=== RUN   TestGetDomainStat_Time_And_Memory
    stats_optimization_test.go:49: time used: 105.315347ms / 300ms
    stats_optimization_test.go:50: memory used: 3Mb / 30Mb
--- PASS: TestGetDomainStat_Time_And_Memory (1.06s)
PASS
ok  	github.com/gt-edu/otus-golang-hw/hw10_program_optimization	1.064s

$ go test -tags bench -v -bench=BenchmarkGetDomainStat -run=^$ -benchmem -benchtime 30s > 30s_readline.out

$ cat 30s_readline.out 
goos: linux
goarch: amd64
pkg: github.com/gt-edu/otus-golang-hw/hw10_program_optimization
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
BenchmarkGetDomainStat
BenchmarkGetDomainStat-4   	51037701	       769.6 ns/op	    4144 B/op	       2 allocs/op
PASS

ok  	github.com/gt-edu/otus-golang-hw/hw10_program_optimization	40.000s

$ benchstat  -delta-test none 30s.out 30s_readline.out
name             old time/op    new time/op    delta
GetDomainStat-4    1.80ms ± 0%    0.00ms ± 0%  -99.96%

name             old alloc/op   new alloc/op   delta
GetDomainStat-4    7.54kB ± 0%    4.14kB ± 0%  -45.00%

name             old allocs/op  new allocs/op  delta
GetDomainStat-4      91.0 ± 0%       2.0 ± 0%  -97.80%
```